refactor component structures
