+++ notes/src/components/NoteDetail/Modal.js
