+++ notes/src/components/NoteDetail/NoteDetail.js
